+++ frontend/catalyst/qjit_device.py
@@ -19,6 +19,40 @@ from pennylane.measurements import MidMeasureMP
 from catalyst.utils.exceptions import CompileError
 from catalyst.utils.patching import Patcher
 
+RUNTIME_OPERATIONS = {
+    "Identity",
+    "PauliX",
+    "PauliY",
+    "PauliZ",
+    "Hadamard",
+    "S",
+    "T",
+    "PhaseShift",
+    "RX",
+    "RY",
+    "RZ",
+    "Rot",
+    "CNOT",
+    "CY",
+    "CZ",
+    "SWAP",
+    "IsingXX",
+    "IsingYY",
+    "IsingXY",
+    "ControlledPhaseShift",
+    "CRX",
+    "CRY",
+    "CRZ",
+    "CRot",
+    "CSWAP",
+    "Toffoli",
+    "MultiRZ",
+    "QubitUnitary",
+    "ISWAP",
+    "PSWAP",
+    "GlobalPhase",
+}
+
 
 class QJITDevice(qml.QubitDevice):
     """QJIT device.
@@ -46,39 +80,7 @@ class QJITDevice(qml.QubitDevice):
     operations = []
     observables = []
 
-    operations_supported_by_QIR_runtime = {
-        "Identity",
-        "PauliX",
-        "PauliY",
-        "PauliZ",
-        "Hadamard",
-        "S",
-        "T",
-        "PhaseShift",
-        "RX",
-        "RY",
-        "RZ",
-        "Rot",
-        "CNOT",
-        "CY",
-        "CZ",
-        "SWAP",
-        "IsingXX",
-        "IsingYY",
-        "IsingXY",
-        "ControlledPhaseShift",
-        "CRX",
-        "CRY",
-        "CRZ",
-        "CRot",
-        "CSWAP",
-        "Toffoli",
-        "MultiRZ",
-        "QubitUnitary",
-        "ISWAP",
-        "PSWAP",
-        "GlobalPhase",
-    }
+    operations_supported_by_QIR_runtime = RUNTIME_OPERATIONS
 
     @staticmethod
     def _get_operations_to_convert_to_matrix(_config):
@@ -218,3 +220,113 @@ class QJITDevice(qml.QubitDevice):
 
         self.check_validity(expanded_tape.operations, [])
         return expanded_tape
+
+
+class QJITDeviceNewAPI(qml.devices.Device):
+    """QJIT device for the new device API.
+    A device that interfaces the compilation pipeline of Pennylane programs.
+    Args:
+        wires (Shots): the number of wires to initialize the device with
+        shots (int): How many times the circuit should be evaluated (or sampled) to estimate
+            the expectation values. Defaults to ``None`` if not specified. Setting
+            to ``None`` results in computing statistics like expectation values and
+            variances analytically
+        backend_name (str): name of the device from the list of supported and compiled backend
+            devices by the runtime
+        backend_kwargs (Dict(str, AnyType)): An optional dictionary of the device specifications
+    """
+
+    # These must be present even if empty.
+    operations = []
+    observables = []
+
+    operations_supported_by_QIR_runtime = RUNTIME_OPERATIONS
+
+    @staticmethod
+    def _check_mid_circuit_measurement(config):
+        return config["compilation"]["mid_circuit_measurement"]
+
+    @staticmethod
+    def _check_adjoint(config):
+        return config["compilation"]["quantum_adjoint"]
+
+    @staticmethod
+    def _check_quantum_control(config):
+        return config["compilation"]["quantum_control"]
+
+    @staticmethod
+    def _set_supported_operations(config):
+        """Override the set of supported operations."""
+        native_gates = set(config["operators"]["gates"][0]["native"])
+        qir_gates = QJITDeviceNewAPI.operations_supported_by_QIR_runtime
+        QJITDeviceNewAPI.operations = list(native_gates.intersection(qir_gates))
+
+        # These are added unconditionally.
+        QJITDeviceNewAPI.operations += ["Cond", "WhileLoop", "ForLoop"]
+
+        if QJITDeviceNewAPI._check_mid_circuit_measurement(config):  # pragma: no branch
+            QJITDeviceNewAPI.operations += ["MidCircuitMeasure"]
+
+        if QJITDeviceNewAPI._check_adjoint(config):
+            QJITDeviceNewAPI.operations += ["Adjoint"]
+
+        if QJITDeviceNewAPI._check_quantum_control(config):  # pragma: nocover
+            # TODO: Once control is added on the frontend.
+            gates_to_be_decomposed_if_controlled = [
+                "Identity",
+                "CNOT",
+                "CY",
+                "CZ",
+                "CSWAP",
+                "CRX",
+                "CRY",
+                "CRZ",
+                "CRot",
+            ]
+            native_controlled_gates = ["ControlledQubitUnitary"] + [
+                f"C({gate})"
+                for gate in native_gates
+                if gate not in gates_to_be_decomposed_if_controlled
+            ]
+            QJITDeviceNewAPI.operations += native_controlled_gates
+
+    @staticmethod
+    def _set_supported_observables(config):
+        """Override the set of supported observables."""
+        QJITDeviceNewAPI.observables = config["operators"]["observables"]
+
+    # pylint: disable=too-many-arguments
+    def __init__(
+        self,
+        original_device,
+        config,
+        backend_name=None,
+        backend_lib=None,
+        backend_kwargs=None,
+    ):
+        self.original_device = original_device
+
+        for key, value in original_device.__dict__.items():
+            self.__setattr__(key, value)
+        self.config = config
+        QJITDeviceNewAPI._set_supported_operations(self.config)
+        QJITDeviceNewAPI._set_supported_observables(self.config)
+
+        self.backend_name = backend_name if backend_name else "default"
+        self.backend_lib = backend_lib if backend_lib else ""
+        self.backend_kwargs = backend_kwargs if backend_kwargs else {}
+
+    def preprocess(
+        self,
+        execution_config: qml.devices.ExecutionConfig = qml.devices.DefaultExecutionConfig,
+    ):
+        """Device preprocessing function."""
+        program, config = self.original_device.preprocess(execution_config)
+        # TODO: Add Catalyst program verification and validation
+        return program, config
+
+    def execute(self, circuits, execution_config):
+        """
+        Raises: RuntimeError
+        """
+        raise RuntimeError("QJIT devices cannot execute tapes.")